requirements.txt
matplotlib==3.3.2
timm==0.6.12
torch==1.13.1
torchvision==0.14.1
typer==0.7.0
umap_learn==0.5.3
tqdm